cleanup page.tsx
